refactor(tools): 進捗表示ヘルパを tools::progress に共通化し book_rescore に進捗を追加
rescore_psv 内に private で埋まっていた進捗表示ヘルパ（compact_count /
compact_rate / fmt_hms / finish_clock / with_progress_keys / MultiFileProgress
(旧 RescoreProgress) / FileProgress）を共有モジュール tools::progress に抽出。
非TTY ログの先頭タグをパラメータ化し、rescore_psv の既存出力文言は不変に保つ。

book_rescore は進捗表示が皆無だったが、go nodes N × 局面数で数時間規模に
なり得るジョブに % / pos/s / 残り時間 / 完了予定時刻を追加（--resume skip 後の
実探索タスク数を分母）。rescore_psv は抽出ブロックを共有モジュール参照へ置換
（挙動不変）。ONNX 専用になった Arc/Mutex/Instant import は cfg gate 化。
